GEODE-4977: Make query results behave like standard containers (#288)
* Adds begin/end to Struct.
* Remove Java style hasNext/next methods.
* Cleanup SelectResults::iterator.
* Remove unused / untested methods.
* Refactor for .NET Enumerable.
* Removes SelectResultsIterator.
* Workaround for Solaris for-range virtual method compiler bug.
* Adds missing Serializable::create(std::basic_string<>)
* Adds missing CacheableKey::create(std::basic_string<>)
* Renames C++ Struct::length to size.
* Renames .NET Struct::Length to Count.

+++ clicache/integration-test/ThinClientRegionQueryTests.cs
@@ -182,7 +182,7 @@ namespace Apache.Geode.Client.UnitTests
 
         ISelectResults<object> results = region.Query<object>(qrystr.Query);
 
-        if (results.Size != QueryStatics.RegionQueryRowCounts[qryIdx])
+        if (results.Size != (ulong) QueryStatics.RegionQueryRowCounts[qryIdx])
         {
           ErrorOccurred = true;
           Util.Log("FAIL: Query # {0} expected result size is {1}, actual is {2}", qryIdx,
